+++ Sources/TuistCache/Utilities/CacheArtifactBuilding.swift
@@ -16,6 +16,7 @@ public protocol CacheArtifactBuilding {
     ///   - deviceName: The specific device that will be used when compiling the given target.
     ///   - into: The directory into which the output artifacts will be copied.
     func build(
+        graph: Graph,
         scheme: Scheme,
         projectTarget: XcodeBuildTarget,
         configuration: String,